Added start/end timestamps
This might fullfill the request (#20) to split the timestamps of the statements that might cover more dates. (need sample data to prove this... since all the current sample data is one day only per statement)

This deprecates the Statement::(set|get)Timestamp methods and data

This deprecates the parse method for retrieving the data too.. SPK and the base engine have been upated to reflect this.

All deprecations will now throw deprecation warnings, but still work using the old way: the statement timestamps were interepereted as Start timestamps. Thus if you have code calling the Statement::Timestamp stuff, prepend this with start :D

This if for feature request done in #20